requirements.txt
flask==2.0.2
py2neo==2021.0.1
numpy==1.19.5
slither-analyzer==0.7.1
nltk==3.6.2
gensim==3.8.1
pandas==1.1.5
scikit-learn==0.24.2
transformers==4.18.0
torch==1.10.1
torchvision==0.11.2
solc-select==0.2.0
solidity-parser==0.0.7
crytic-compile==0.1.13